move script init to App.tsx
