+++ client/one-time-payment/react/src/App.tsx
@@ -1,7 +1,29 @@
+import { useState, useEffect } from "react";
 import PayPalButton from "./components/PayPalButton";
 import VenmoButton from "./components/VenmoButton";
 
 function App() {
+  const [isSDKReady, setIsSDKReady] = useState<boolean>(false);
+
+  useEffect(() => {
+    const loadPayPalSDK = async () => {
+      try {
+        if (!window.paypal) {
+          const script = document.createElement("script");
+          script.src = "https://www.sandbox.paypal.com/web-sdk/v6/core";
+          script.async = true;
+          script.onload = () => setIsSDKReady(true);
+          document.body.appendChild(script);
+        } else {
+          setIsSDKReady(true);
+        }
+      } catch (e) {
+        console.error("Failed to load PayPal SDK", e);
+      }
+    };
+
+    loadPayPalSDK();
+  }, []);
   return (
     <>
       <h1>React One-Time Payment Recommended Integration</h1>
@@ -12,8 +34,8 @@ function App() {
           gap: "12px",
         }}
       >
-        <PayPalButton />
-        <VenmoButton />
+        {isSDKReady && <PayPalButton />}
+        {isSDKReady && <VenmoButton />}
       </div>
     </>
   );
